+++ projects/ngx-action/src/lib/internals/symbols.ts
@@ -1,5 +1,3 @@
 export const isInitCalledSymbol = Symbol('isInitCalled');
-export const isDecoratorAppliedSymbol = Symbol('isDecoratorAppliedSymbol');
 export const actionHandlersSymbol = Symbol('actionHandlersSymbol');
 export const asyncActionHandlersSymbol = Symbol('asyncActionHandlersSymbol');
-export const destroySubjectSymbol = Symbol('destroySubjectSymbol');