+++ projects/ngx-action/src/public-api.ts
@@ -12,4 +12,3 @@ export * from './lib/actions.module';
 export * from './lib/actions';
 export * from './lib/async-action-handler';
 export * from './lib/init-action-handlers';
-export * from './lib/with-action-handlers';