+++ CHANGELOG.md
@@ -1,4 +1,12 @@
+## 2.0.0
+
+### Breaking changes
+
+- Angular 16
+- Remove `WithActionHandlers` decorator
+- `initActionHandlers(this)` should be called inside class constructor
+
 ## 1.1.0
 
 - Add `ActionClass` type to public api
-- Allow to dispatch multiple actions and subscribe to multiple actions.
+- Allow to dispatch multiple actions and subscribe to multiple actions